Make do_path_regex_sub() honour string regex arguments
The Plugin.do_path_regex_sub() method to apply regex substitutions
to paths matching a pattern documents that it accepts either a
compiled re object or a regular expression as a string:

    '''Apply a regexp substituation to a set of files archived by
    sos. The set of files to be substituted is generated by matching
    collected file pathnames against pathexp which may be a regular
    expression string or compiled re object. The portion of the file
    to be replaced is specified via regexp and the replacement string
    is passed in subst.'''

It lies. Attempting to pass a string for the 'pathexp' parameter
will result in:

    Traceback (most recent call last):
      File "/usr/sbin/sosreport", line 23, in <module>
        main(sys.argv[1:])
      File "/usr/lib/python2.6/site-packages/sos/sosreport.py", line 1229, in main
        sos.execute()
    AttributeError: 'str' object has no attribute 'match'

    > /usr/lib/python2.6/site-packages/sos/plugins/__init__.py(219)do_path_regex_sub()
    -> match = pathexp.match

Look to see if the object we are passed has a 'match()' method and
call re.compile on it if it does not.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>

Conflicts:
	sos/plugins/__init__.py